+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 # 0.0.1
 ## Added
 - SDK released
+# 0.0.2
+## Added
+- Support for "genesis ID" field in transactions
+- Support for "close remainder to" field in transactions